Cache spec/products in-isolate; inline executors into tools; delete executor.ts
Performance:
- Add src/spec-cache.ts: module-scope cache (1h TTL) for spec.json and
  products.json. The MCP isolate stays warm across requests, so this removes
  an R2 round-trip from every search/execute/non-codemode call. search/
  non-codemode and the search-tool description now read from the cache.

Investigated reusing the Worker Loader isolate (the crypto.randomUUID
workerId): NOT possible. Dynamic-worker isolates hard-disable eval/new
Function ('Code generation from strings disallowed', confirmed even with
compat flags), so code that varies per call cannot be passed into a warm
isolate — it must be baked into the module source, one module per distinct
code. This matches the Code Mode docs ('each execution gets its own Worker
instance'). Per-user warm execute is therefore impossible and also unsafe
(execute runs mutations; blind retry would double-apply). Left as-is.

Structure:
- Delete src/executor.ts; inline createCodeExecutor into tools/execute.ts
  (runExecute) and createSearchExecutor into tools/search.ts (runSearch).
  There was no shared logic worth a module — the two isolates differ in
  globalOutbound, bound module source, and entrypoint.

Tests:
- New tests/helpers/spec.ts owns the test-only spec fixtures (seedSpec /
  clearSpec), which also reset the in-isolate cache. Test files import these
  instead of putting fixture logic inline or reaching into src/. The cache
  primitive resetSpecCache stays in src/spec-cache.ts and is used only by
  the helper.
